Initial work on Table of Contents generator
Plan: writing the Table of Contents generator as a series of
Jinja filters that must be used together.

Benefits: allow us to place the Table of Contents anywhere in the page,
as this is an extension at the Pelican level.
Writing a simple Markdown postprocessor would limit us in terms of the
location of the Table of Contents: it would need to be located within the
page content's scope.