Updating navbar headers as per mizar's instructions
Currently buddycloud.com is fully static, so the header for any buddycloud.com pages doesn't update to show the logged person, after they have logged into hosting.buddycloud.com (or .net).
An integration between this and the hosting project needs to be done if this is the desired effect (and it seems like it is).